+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Center map when orderForm address has an empty geoCoordinate array.
+
 ## [0.11.1] - 2023-11-07
 
 ### Fixed